text-pipeline/requirements.txt
# OCR / Image
Pillow
opencv-python
numpy
paddlepaddle
paddleocr
pytesseract

# Text preprocessing
regex
langdetect
wordfreq

rapidfuzz
symspellpy

# Config
PyYAML

# Infra / schema
pydantic
python-dotenv

rich

# pip install -r requirements.txt
